Initializing all the bulk information with the constructor so it doesnt need to be created every instance. Ask for movie choice is repeately used for every option in which a parameter, movies, is printed out for the user and then processed through the users choice for the specified movie information. This is used by find cast, find genre etc. For ratings and revenue, no alphabetical sorting is used which is highlighed by askForMovieNoAlphabetical.
The methods such as find cast and find genre are more or less the same with different parameters and used variables since they require the same process but different values.

The thinking for the top50 items was to have a method find greatest to go through all the movies and find the one with the greatest respective category (ex: score). Then in another method, using a for loop 50 times, we find the greaest 50 times (removing the previous each time). Then we have the list of greastest movies we can put it through askForMovieNoAlphabetical.